@@ -8,43 +8,31 @@ description: >
 
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  analyzer: ">=0.39.0 <0.42.0"
-  args: ^1.5.1
-  codemod: ^0.3.0
-  collection: ^1.14.7
-  glob: ^1.2.0
-  logging: ^0.11.3+2
-  meta: ">=1.2.2 <1.7.0" # Workaround to avoid https://github.com/dart-lang/sdk/issues/46142
-  path: ^1.6.2
-  prompts: ^1.3.1
-  pub_semver: ^1.4.2
-  source_span: ^1.4.1
-  yaml: ^2.1.16
+  analyzer: ^1.7.2
+  args: ^2.2.0
+  codemod: ^1.0.1
+  collection: ^1.15.0
+  glob: ^2.0.1
+  logging: ^1.0.1
+  meta: ^1.7.0
+  path: ^1.8.0
+  prompts: ^2.0.0
+  pub_semver: ^2.0.0
+  source_span: ^1.8.1
+  yaml: ^3.1.0
 
 dev_dependencies:
-  build_runner: ^1.0.0
-  build_test: ">=0.10.9 <2.0.0"
-  dart_dev: ^2.0.1
-  dart_style: ^1.3.3
-  dependency_validator: ^1.2.3
-  mockito: ^4.0.0
-  react: ^4.6.1
-  test: ^1.15.7
-  test_descriptor: ^1.2.0
-  workiva_analysis_options: ^1.1.0
+  build_runner: ^2.1.1
+  build_test: ^2.1.3
+  dart_style: ^2.0.3
+  dependency_validator: ^3.1.0
+  mockito: ^5.0.14
+  test: ^1.17.10
+  test_descriptor: ^2.0.0
+  workiva_analysis_options: ^1.2.2
 
 executables:
-  boilerplate_upgrade:
-  component2_upgrade:
-  dart2_upgrade:
-  react16_ci_precheck:
-  react16_consumer_overlay_update:
-  react16_dependency_override_update:
-  react16_post_rollout_cleanup:
-  react16_upgrade:
-  react17_dependency_override_update:
-  react17_upgrade:
   dart2_9_upgrade:
