@@ -3,6 +3,7 @@
 
 import 'package:analyzer/dart/ast/ast.dart';
 import 'package:analyzer/dart/ast/visitor.dart';
+import 'package:analyzer/dart/element/type.dart';
 
 /// A usage of an OverReact component via its fluent interface.
 class FluentComponentUsage {
@@ -12,7 +13,7 @@ class FluentComponentUsage {
   final InvocationExpression node;
 
   /// The cascade of this usage (unwrapped from parens), or `null` if it doesn't have one.
-  final CascadeExpression cascadeExpression;
+  final CascadeExpression? cascadeExpression;
 
   /// The expression upon which the cascade is performed,
   /// and that is invoked with children to build the component.
@@ -22,11 +23,12 @@ class FluentComponentUsage {
   /// Usually a [MethodInvocation] or [Identifier].
   final Expression builder;
 
-  String get componentName => getComponentName(builder);
+  String? get componentName => getComponentName(builder);
 
-  bool get isDom =>
-      const ['DomProps', 'SvgProps'].contains(builder.staticType?.name);
-  bool get isSvg => const ['SvgProps'].contains(builder.staticType?.name);
+  bool get isDom => const ['DomProps', 'SvgProps']
+      .contains(builder.staticType?.interfaceTypeName);
+  bool get isSvg =>
+      const ['SvgProps'].contains(builder.staticType?.interfaceTypeName);
 
   /// Whether the invocation contains one or more children passed as arguments instead of a list.
   bool get hasVariadicChildren =>
@@ -35,6 +37,17 @@ class FluentComponentUsage {
 
   /// The number of child arguments passed into the invocation.
   int get childArgumentCount => node.argumentList.arguments.length;
+}
+
+extension _TryCast<T> on T {
+  S? tryCast<S extends T>() {
+    final self = this;
+    return self is S ? self : null;
+  }
+}
+
+extension on DartType? {
+  String? get interfaceTypeName => tryCast<InterfaceType>()?.element.name;
 }
 
 /// Returns the OverReact fluent interface component for the invocation expression [node],
@@ -52,11 +65,11 @@ class FluentComponentUsage {
 ///     * e.g., `getButtonBuilder()`
 /// * `*builder*`
 ///     * e.g., `var buttonBuilder = Button();`
-FluentComponentUsage getComponentUsage(InvocationExpression node) {
+FluentComponentUsage? getComponentUsage(InvocationExpression node) {
   var functionExpression = node.function;
 
   Expression builder;
-  CascadeExpression cascadeExpression;
+  CascadeExpression? cascadeExpression;
 
   if (functionExpression is ParenthesizedExpression) {
     var expression = functionExpression.expression;
@@ -73,13 +86,14 @@ FluentComponentUsage getComponentUsage(InvocationExpression node) {
   bool isComponent;
   if (builder.staticType != null) {
     // Resolved AST
-    isComponent = builder.staticType.name?.endsWith('Props') ?? false;
+    isComponent =
+        builder.staticType.interfaceTypeName?.endsWith('Props') ?? false;
   } else {
     // Unresolved AST (or type wasn't available)
     isComponent = false;
 
     if (builder is MethodInvocation) {
-      String builderName = getComponentName(builder);
+      String? builderName = getComponentName(builder);
 
       if (builderName != null) {
         isComponent =
@@ -99,10 +113,10 @@ FluentComponentUsage getComponentUsage(InvocationExpression node) {
   return FluentComponentUsage._(node, cascadeExpression, builder);
 }
 
-String getComponentName(Expression builder) {
+String? getComponentName(Expression builder) {
   if (builder.staticType != null) {
     // Resolved AST
-    final typeName = builder.staticType?.name;
+    final typeName = builder.staticType?.interfaceTypeName;
     if (typeName == null) return null;
     if (const ['dynamic', 'UiProps'].contains(typeName)) return null;
     if (builder is MethodInvocation) {
@@ -117,7 +131,7 @@ String getComponentName(Expression builder) {
     // Unresolved
     String builderName;
     if (builder.target != null) {
-      builderName = builder.target.toSource() + '.' + builder.methodName.name;
+      builderName = builder.target!.toSource() + '.' + builder.methodName.name;
     } else {
       builderName = builder.methodName.name;
     }
@@ -169,7 +183,7 @@ bool hasChildComponent(ArgumentList arguments) {
 
 /// Attempt to find and return the closest expression that encloses the [node]
 /// and is an independent Flutter `Widget`.  Return `null` if nothing found.
-FluentComponentUsage identifyUsage(AstNode node) {
+FluentComponentUsage? identifyUsage(AstNode? node) {
   for (; node != null; node = node.parent) {
     if (node is InvocationExpression) {
       final usage = getComponentUsage(node);
