@@ -19,255 +19,6 @@ import 'package:test/test.dart';
 
 void main() {
   group('Boilerplate Utilities', () {
-    group('isPropsUsageSimple', () {
-      int assertionCount;
-
-      setUp(() {
-        assertionCount = 0;
-      });
-
-      test('will throw if not passed a props class', () {
-        final input = '''
-            @Factory()
-            UiFactory<FooProps> Foo =
-                // ignore: undefined_identifier
-                \$Foo;
-
-            @Props()
-            class _\$FooProps extends UiProps{
-              String foo;
-              int bar;
-            }
-
-            @Component2()
-            class FooComponent extends UiComponent2<FooProps> {
-              @override
-              render() {
-                return Dom.ul()(
-                  Dom.li()('Foo: ', props.foo),
-                  Dom.li()('Bar: ', props.bar),
-                );
-              }
-            }
-          ''';
-
-        final unit = parseString(content: input).unit;
-        final classDeclarations =
-            unit.declarations.whereType<ClassDeclaration>();
-
-        expect(classDeclarations.length, 2);
-
-        classDeclarations.forEach((classNode) {
-          if (!extendsFromUiPropsOrUiState(classNode)) {
-            assertionCount++;
-            expect(() => isSimplePropsOrStateClass(classNode),
-                throwsA(TypeMatcher<AssertionError>()));
-          }
-        });
-
-        expect(assertionCount, 1);
-      });
-
-      group('returns true if', () {
-        test(
-            'there are no mixins and the class extends from UiProps or UiState',
-            () {
-          final input = '''
-            @Factory()
-            UiFactory<FooProps> Foo =
-                // ignore: undefined_identifier
-                \$Foo;
-
-            @Props()
-            class _\$FooProps extends UiProps {
-              String foo;
-              int bar;
-            }
-
-            @State()
-            class _\$FooState extends UiState {
-              String foo;
-              int bar;
-            }
-
-            @Component2()
-            class FooComponent extends UiComponent2<FooProps> {
-              @override
-              render() {
-                return Dom.ul()(
-                  Dom.li()('Foo: ', props.foo),
-                  Dom.li()('Bar: ', props.bar),
-                );
-              }
-            }
-          ''';
-
-          final unit = parseString(content: input).unit;
-          final classDeclarations =
-              unit.declarations.whereType<ClassDeclaration>();
-
-          expect(classDeclarations.length, 3);
-
-          classDeclarations.forEach((classNode) {
-            if (isAPropsOrStateClass(classNode)) {
-              assertionCount++;
-              expect(isSimplePropsOrStateClass(classNode), isTrue);
-            }
-          });
-
-          expect(assertionCount, 2);
-        });
-      });
-
-      group('returns false if', () {
-        test(
-            'there are mixins and the class doesn\'t extend from UiProps or UiState',
-            () {
-          final input = '''
-            @Factory()
-            UiFactory<FooProps> Foo =
-                // ignore: undefined_identifier
-                \$Foo;
-
-            @Props()
-            class _\$FooProps extends ADifferentPropsClass with APropsMixin {
-              String foo;
-              int bar;
-            }
-
-            @State()
-            class _\$FooState extends ADifferentStateClass with AStateMixin {
-              String foo;
-              int bar;
-            }
-
-            @Component2()
-            class FooComponent extends UiComponent2<FooProps> {
-              @override
-              render() {
-                return Dom.ul()(
-                  Dom.li()('Foo: ', props.foo),
-                  Dom.li()('Bar: ', props.bar),
-                );
-              }
-            }
-          ''';
-
-          final unit = parseString(content: input).unit;
-          final classDeclarations =
-              unit.declarations.whereType<ClassDeclaration>();
-
-          expect(classDeclarations.length, 3);
-
-          classDeclarations.forEach((classNode) {
-            if (isAPropsOrStateClass(classNode)) {
-              assertionCount++;
-              expect(isSimplePropsOrStateClass(classNode), isFalse);
-            }
-          });
-
-          expect(assertionCount, 2);
-        });
-
-        test('there are mixins but the class extends from UiProps or UiState',
-            () {
-          final input = '''
-            @Factory()
-            UiFactory<FooProps> Foo =
-                // ignore: undefined_identifier
-                \$Foo;
-
-            @Props()
-            class _\$FooProps extends UiProps with APropsMixin {
-              String foo;
-              int bar;
-            }
-
-            @State()
-            class _\$FooState extends UiState with AStateMixin {
-              String foo;
-              int bar;
-            }
-
-            @Component2()
-            class FooComponent extends UiComponent2<FooProps> {
-              @override
-              render() {
-                return Dom.ul()(
-                  Dom.li()('Foo: ', props.foo),
-                  Dom.li()('Bar: ', props.bar),
-                );
-              }
-            }
-          ''';
-
-          final unit = parseString(content: input).unit;
-          final classDeclarations =
-              unit.declarations.whereType<ClassDeclaration>();
-
-          expect(classDeclarations.length, 3);
-
-          classDeclarations.forEach((classNode) {
-            if (isAPropsOrStateClass(classNode)) {
-              assertionCount++;
-              expect(isSimplePropsOrStateClass(classNode), isFalse);
-            }
-          });
-
-          expect(assertionCount, 2);
-        });
-
-        test(
-            'there are no mixins but the class doesn\'t extend from UiProps or UiState',
-            () {
-          final input = '''
-            @Factory()
-            UiFactory<FooProps> Foo =
-                // ignore: undefined_identifier
-                \$Foo;
-
-            @Props()
-            class _\$FooProps extends ADifferentPropsClass {
-              String foo;
-              int bar;
-            }
-
-            @State()
-            class _\$FooState extends ADifferentStateClass {
-              String foo;
-              int bar;
-            }
-
-            @Component2()
-            class FooComponent extends UiComponent2<FooProps> {
-              @override
-              render() {
-                return Dom.ul()(
-                  Dom.li()('Foo: ', props.foo),
-                  Dom.li()('Bar: ', props.bar),
-                );
-              }
-            }
-          ''';
-
-          final unit = parseString(content: input).unit;
-          final classDeclarations =
-              unit.declarations.whereType<ClassDeclaration>();
-
-          expect(classDeclarations.length, 3);
-
-          classDeclarations.forEach((classNode) {
-            if (isAPropsOrStateClass(classNode)) {
-              assertionCount++;
-              expect(isSimplePropsOrStateClass(classNode), isFalse);
-            }
-          });
-
-          expect(assertionCount, 2);
-        });
-      });
-    });
-
     group('getSemverHelper() with isPublic() and getPublicExportLocations()',
         () {
       group('with --treat-all-components-as-private flag', () {
@@ -303,7 +54,7 @@ void semverUtilitiesTestHelper({
   bool shouldTreatAllComponentsAsPrivate = false,
   bool isValidFilePath = true,
 }) {
-  SemverHelper semverHelper;
+  late SemverHelper semverHelper;
 
   setUpAll(() {
     semverHelper = getSemverHelper(path,
