@@ -34,11 +34,11 @@ class GeneratedFactoryMigrator extends RecursiveAstVisitor
       final generatedFactory = getGeneratedFactory(node);
       final parentMethod = generatedFactory?.parent?.parent;
       if (parentMethod is MethodInvocation &&
-          parentMethod.methodName?.name == castFunctionName) {
+          parentMethod.methodName.name == castFunctionName) {
         return;
       }
 
-      yieldPatch('$castFunctionName(', generatedFactory.offset,
+      yieldPatch('$castFunctionName(', generatedFactory!.offset,
           generatedFactory.offset);
       yieldPatch(')', generatedFactory.end, generatedFactory.end);
     }
