@@ -56,7 +56,7 @@ class LatestBoilerplateVisitor extends RecursiveAstVisitor
       final generatedFactory = getGeneratedFactory(node);
       final parentMethod = generatedFactory?.parent?.parent;
       if (parentMethod is MethodInvocation &&
-          parentMethod.methodName?.name == castFunctionName) {
+          parentMethod.methodName.name == castFunctionName) {
         detectedLatestBoilerplate = true;
       }
     }
